Convert current @property decorators to cached where possible
Many of current properties can be cached - their values will not change
during the citool's session, and it makes more sense e.g. to create task
name just once, and return the cached value. We did this for some
properties already but it would be too laborious to do it the same way
for all such properties, therefore I decided to give it nicer & easier
to use form.