Adds tests for more CI methods
 - CI._check_module_file
 - deleting shared functions

Also, NonLoadingCI class moved into __init__.py, it will be used in
tests covering modules so it must be available for them as well.

And removing test_import_sanity which simply instantiates CI - this test
hides the fact that most of the code is *not* covered by any other test
than simple "run it & watch it burn".