Quote command-line options when logging current command line
Apparently, quotes are necessary, otherwise more complex option values
would be split up.

  `citool foo -a -b "--bar --baz"`

is logged as

  `citool \
      foo -a -b --bar --baz`

which is not what we want :/ With this patch we get:

  `citool \
      foo "-a" "-b" "--bar --baz"`

When pasted back to terminal, shell will take care of quotes, and it
will work as expected. Well, I hope it will, my shell-fu is not as
mighty as my python-fu :/