Allow "soft" exceptions
Soft exception is the exception we, as in developers and operations
team, don't have to be aware. There are several examples of such
exceptions already in the Jenkins jobs, e.g. component general test plan
without any relevant tests is such an event - it's an unrecoverable
error, citool has nothing to do anymore, and we must inform the issuer,
but it's not necessary to drag CI stuff into the notifications.
Exception is raised in such case, and notification modules should react
appropriently, e.g. by not sending mails to qe-baseos-automation.