Handle SIGTERM just like SIGINT (or Ctrl+C)
Ctrl+C is translated to KeyboardInterrupt exception, which citool
handles. However Jenkins, when terminating job - or any other sane
process management system - uses SIGTERM which was not handled by citool
at all. Python's pre-installed SIGINT handler is fine, it translated
signal to an exception, so lets reuse it for SIGTERM as well, just add
some logging for us, and I think we're done.