Implement basic filtering in brew-dispatch configuration
See brew-dispatch sources for more details, but in short:

 - list commands for your component like usually:

    component:
        - command1
        - command2

 - or add commands for specific builds:

    component:
        extra-testing:
            - BREW_TASK_TARGET == 'rhel-7.4-candidate'
            - command1
            - command2
        extra-special-testing:
            - BREW_TASK_TARGET == 'rhel-6.5-z-candidate'
            - command3
            - command4
        default:
            - command5

This allows more fine-grained filtering of commands. This way, when brew
build's target is `'rhel-7.4-candidate'`, `command1` and `command2` will
be dispatched, `command3` and `command4` will be dispatched when target
is `'rhel-6.5-z-candidate'`, and for every other build, `command5` will
run.