Clear Jenkins workspace before starting builders
Uses Workspace Cleanup Plugin which is available to us. Very useful when
we start with Restraint since it tends to put its state data into
temporary directories, and we don't remove them because we don't care
about the directory name, we are happy with its content being available
to people who search for logs. Which is good, but the next job will
store its data into a different temporary dir, and none of them gets
removed...