notify-email: use dummy task when no real is available
This can happen - build-on-commit success will lead to notify-email -
which is there only to send e-mail on failures - and its execute()
expects brew task which is not provided by any module in the pipeline.
Therefore successfull build-on-commit crashes CI pipeline, and sends
stupid e-mail. This patch uses dummy brew task, previously used by
destroy() method, even for execute(), and emits a warning.