notify-recipients: support multiple --*-*-notify options
Only a single instance of, for example, --wow-add-notify was allowed.
There not really a reason why to disallow multiple uses of these
options. When it comes to configuration file, still only one can be
used but those passed on the command-line will override it anyway so
there's no conflict.

Implements piece necessary for PLATFORMCI-1469.