koji: add new module for Fedora's koji
This module adds koji module. The module file
is named koji_fedora to mitigate collision with koji
in pytest.

This module forms basis for the new brew module, which
will extend KojiTask class for RHEL specific functionality.

Add wait option which causes koji to wait until a task is finished.
This is required for koji dispatcher, which reacts on completed
koji build, but it takes sometimes a bit of additiona time until
the task is properly finished.

This patch also adds possibility to resolve build by:

    - build_id
    - nvr
    - latest build of given package with a specific tag

Tests for the module are included in tests/test_koji.py test.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>