Testing thread ID
We lack some hierarchy when dealing with multiple Jenkins builds started
by a single brew-dispatcher job. This patch tries to introduce some
order into the flock of jobs.

brew-dispatcher, or any other job that's at the beginning of testing
pipeline, generates a "testing thread ID" for each build it dispatches.
It can be a single build (e.g. mtf which is triggered by CI bus *and*
does the testing itself as well) or multiple builds (brew-dispatcher).

This ID remains the same (unless user deliberately changes it) over
the Jenkins build's rebuilds, and we can easily say "this build was
started by this build" simply by looking on its name, or "build C has
the same thread ID as build B, which is older, therefore C is rebuild of
B".

Simple approach to thread ID creation is to simply concatenate IDs over
the flow of dispatching builds. So far, we have at max. two levels of
builds - ci-brew-dispatcher vs builds it dispatcher (ci-beaker, ci-rpmdiff,
...). ci-brew-dispatcher get the initial thread ID (it starts the
hierarchy), then simply adds and index of the dispatcher build, and
tells the dispatched build to use this new value as its thread ID. When
this build of 2nd generation wants to dispatch builds of its own, it can
use the same process, deepening the "tree" of thread IDs.