openstack: when hitting quotas, wait for OpenStack resources instead of quiting
We used to raise an exception when we got an exception from OpenStack.
This patch handles one specific instance raised when *.create operation
would overcome the quota. In such case, we wait and try again later.

At this moment, there is not specific timeout for the wait but it won't
run forever, the pipeline will get killed by our CI watchdog.