beaker, restraint-scheduler: use --install-method to override brew-build's METHOD variable
We need this to cleanly support RPM_BLACKLIST feature of brew-build.

Note the different default install method of beaker and
restraint-scheduler (localupdate vs localinstall). This is a historic
weirdness which I'd like to unify some day in the future but right now,
I don't have idea what I might break by changing that, therefore I'm not
touching it.

Necessary for PLATFORMCI-1934.