jenkins: disable default generation of JJB config file
'jenkins' module is not the correct place for this action, especially
when we moved from bothering citool pipelines with creating their Jenkins
jobs. And it even breaks use of gluetool/citool in Docker when the
pipeline runs under non-root account inside the container. As I don't
have any better place so far, I simply disable the automatic generation
of JJB's config file - if you need it, you can always run the module
with --create-jjb-config option.