wow: generalized option map, moving large part of wow options into configuration
The primary goal was to open wow module to deal with requirements,
raised recently due to state of Beaker resources and their usage. We had
to skip s390x boxes altogether, and we must somehow deal with
RHEL/RHEL-ALT inheritance, where each distro family might require
different wow options to be used to actually obtain the desired Beaker
job. Wow options mapping was natural but proved to be not flexible
enough.

Therefore, this patch moves large part of wow options out of the code,
the logic now lies in the configuration file. Equiped with rules and
templates, seems to be a good start, allowing for paralelization on
wow/beaker level.

Example of a wow options map:

// Default options, commong for all cases
- rule: BUILD_TARGET.match('.*')
  add-options: |
    --distro "{{ DISTRO }}"
    --taskparam "BASEOS_CI=true"
    --taskparam "BEAKERLIB_RPM_DOWNLOAD_METHODS=yum\\ direct"
    --taskparam "BASEOS_CI_TASKS={{ TASKS | map(attribute='task_id') | map('string') | join(',') }}"
    --taskparam "BASEOS_CI_COMPONENT={{ PRIMARY_TASK.component  }}"
    --taskparam "BASEOS_CI_TASK_TARGET={{ PRIMARY_TASK.target }}"
    --setup beakerlib

// Avoid s390x everywhere
- rule: BUILD_TARGET.match('.*')
  add-options: --no-arch s390x

// RHEL-7.5 builds are tested on RHEL-ALT but avoid RHEL-ALT @ ppc64le - it'd require Power9
- rule: BUILD_TARGET.match('rhel-7\..(?:-[zZ])?-candidate') and DISTRO.match('^rhel-alt-7\.5.*')
  add-options: --no-arch ppc64le

guess-beaker-distro may provide multiple distros for a given target - to
deal with such situation, wow module simply constructs a separate job
for each distro, and it's up to wow's user (beaker/restraint-*) modules
to merge them. If guess-beaker-distro returns two distros for rhel-7
target (e.g. rhel-7.5, rhel-alt-7.5), 'ppc64le' rule would match only
for rhel-alt-7.5 (because of the DISTRO match) - rhel-7.5 *would be*
scheduled on ppc64le.

Beaker module merges multiple wow jobs by giving their IDs to beaker-jobwatch,
which merges them into a single matrix URL, while restraint-* modules just
raise the exception - this kind of scenarios is not yet required in OpenStack
testing. Should users ask for this feature, it may be implemented there as
well (but there's just a single arch in OpenStack so far, so...)

Rules and their evaluation is implemented in a fresh new module,
rule-engine. It is based on code in test-batch-planner module, which, in
a separate patch, will be removed in favor of rule-engine. That will
require a change of test-batch-planner configuration (brew-dispatcher.yaml).