restraint-runner: allow continuation of a test run after restraint's error
By default, an error reported by restraint leads to en immediate end of
the whole task set. Some QEs are interested in getting results for the
remaining test cases as well, this patch adds necessary options. User
can now instruct restraint-runner to store the snapshot and to move on
to the next task. Obviously, the use of such feature with snapshots
disabled would be negligible, however the patch cleans a code a bit, so
there's a profit for common user as well.

Implements PLATFORMCI-2052.