openstack: shutdown an instance, and wait for it
There was a missing bit, there was no "do shotdown" action in `shutdown`
method :( The patch fixes that, and adds a wait for the desired state.
That led me to refactoring our "wait for status" code, which is now
simplified and does not use `retry` anymore - when we wait, we simply
use glouetool's `wait` method.