Rename brew-build-name to jenkins-build-name
Brew-build-name is not actually related to Brew any more. It is generic
jenkins build name setter, so it is valid to rename it.

Also copr-build-name is added to names, to use this module in copr dispatcher
jenkins job.

Name is passed as a param. It can contains keywords from artifacts context
(shared function `artifact_context` of brew/koji/copr module). These keywords
are substitute by their values.