openstack: with snapshots enabled, rebuild may fail because of ineffective shutdown
Scene: test fails to finish and leaves the guest in absolutely broken
state. Test runner wishes to move on to the next test, and asks guest to
be rebuilt from a snapshot. First, instance must be stopped, therefore
its stop() method is called, followed by waiting for the instance to get
into SHUTOFF state - and this fails if the software stack on the guest
was broken enough to be unable to actually shutdown the guest. stop()
tried to "gracefully" shutdown the instance, asking its OS to halt, and
this may be impossible, therefore the wait fails as the instance remains
ACTIVE.

There's no "force power off" method in Nova client API, therefore it's
not possible to break this cycle. Guest might try a hard reboot, but
that quite easily leads to broken system again. To deal with this loop,
guest releases the instance back to OS pool, and gets a new one using
the snapshot as its image. Floating IP is re-assigned to the new
instance, and the guest may return back to the test runner like nothing
happened - it is backed by an instance running requested snapshot.

To do this, the patch splits OpenStack guest's __init__ to smaller
steps in _acquire_* methods, providing _release_* where necessary, and
extending rebuild from snapshot to catch shutdown timeouts, falling back
to provisioning new instance.

The rest of the patch is just a shuffling and related cleanup of OS
guest methods, to sort them a bit in order from lower-levels to higher
ones, renaming variables holding low-level references to OS resources to
reflect this relationship.