tcms-run: create a TCMS run, and fill it with data from Beaker jobs
Use case: BaseOS QE Tools want to use TCMS for results of some of our
components in CI. TO do that, adding this module. Still there's a lot to
stablize, like how the runs should be named, or whether to recycle them
and so on, but it's a working code, leaving lot to configuration and
user's demands.

And one day we're gonna cut this in half and use it for EWA :)