koji: provide list of arches the task was built for
It is either directly in the task info, if it was built for limited set
of arches, or we inspect its buildArch children for complete set.

restraint-scheduler immediately makes use of this information byt
raising error if the task does not contain x86_64 artifacts, as that's
so far only supported ach in our OpenStack testing. Not happy about it,
restraint-scheduler is not freely usable with e.g. Beaker boxes :/ But
that's for another patch since there's already '--arch=x86_64' hardcoded
in its code.

beaker module is simple, it's going to make use of this bit via wow
module:

    {% set task_arches = PRIMARY_TASK.task_arches %}
    {% if not task_arches.complete %}
        {% for arch in task_arches.arches %}
            --arch={{ arch }}
        {% endfor %}
    {% endif %}

If the list was complete (whatever that means within the context of
Koji/Brew setup, tags, targets, etc.), we don't apply any additional
options. If the list of arches *was* limited, let's ask wow to schedule
only for those listed.

It's not bullet-proof, we may end-up with the empty XML anyway, for
example task for ppc (no ppc64 nor ppc64le, just ppc, secondary arch)
would probably wouldn't have any relevant tests, depending on component,
or even a usable distro... But I don't see any easy solution to that,
module's would have to know what arches are allowed for each distro and
target and so on, and do some filtering... :/ I suspect there won't be
many of these weird tasks, in BaseOS Tools QE I saw like on eso far.