pipeline-state-reporter: allow details of other types, not just strings
All artifact details are rendered as templates, therefore every detail
is a string. But, some details are expected to have other type, e.g.
branch can be either string or null, or scratch which is always boolean.
That cannot be done via templating (always returns a string), therefore
adding an extra flag to evaluate such set as rules, that yields other
Python types easily. But user has to ask for that explicitly.