beaker: add support for SUT setup failures
This patch provides a possiblity to specify a list of SUT installation
tasks, which cause the module to throw a SUTInstallationFailedError
exception, similarly as openstack testing does.

Also fix issue with reusing jobs.

Changelog: SUT installation errors in Beaker jobs are now detected and properly reported instead of looking as a test failure.

Resolves PLATFORMCI-2018

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>