brew-build-task-params: control of /distribution/install/brew-build params
This has been present in two instances, one for OpenStack branch, one
for Beaker-based testing, as both call this task in a different way. But
the params being used are still the same, and it has bitten us finally
when beaker module does not provide correct SERVER variable.

This module takes care of creating options for this task, and it's left
up to its users to pass them to the task.