pipeline-state-reporter: add a timestamp the message was generated at
Several sources can be used to observe UMB traffic, and even examine
gathered messages. However, there seem to be delays, and I find the need
for the accurate timestamp, representing the time when the message was
born, to have something to compare with other timestamps in play. It
also helps to order messages, when we can use our UTC timestamp instead
of suspicious "timestamp" field of captured messages which may represent
other moments, e.g. the moment the message was captured by Datagrepper
(few hours after it was generated...).