koji: refactored to provide task(s) even if they don't have any artifacts
Raise an exception but provide access to the tasks via primary_task and
tasks shared functions after that. This should solve issues for
following modules that would like to e.g. notify user but has no access
to the artifact, since koji module gave up. After this patch, exception
was raised, therefore pipeline is going to crash, but notifications can
work with an artifact representation.

Included is a cleanup of the code that's touching Koji API, since we
must be a bit smarter when guessing SRPM name. Also the tests have
changed, moving the mock API responses to YAML files under assets
directory.