restraint-scheduler: raise NoTestableArtifactsError instead of generic error
Raised when task has no artifacts compatible with OpenStack's
x86_64/noarch environment. Raise specific exception instead of a generic
one to allow its filtering in pipeline state reporter and other modules.

There is nothing to fix, this is not a bug or a temporary infrastructure
issue, it's just a design and/or limitation we have to live with.
Therefore it's be nice to give it nicer e-mails and more reasonable
pipeline final state.