restraint-scheduler: use --no-arch to limit environment
To limit to just supported architectures, using --arch=foo would work fine
until the testing runs into an artifact with incomplete set of arches, with
foo present. Configuration would try to limit recipe sets to just those arches
present, add --arch=foo. THis scheduler would try to limit arches even more,
to supported ones only, adding another --arch=foo, which would make wow construct
*two* same recipeSets for arch foo, possibly leading to provisioning two boxes
for this arch, running the exactly same set of tasks.

On the other hand, multiple --no-arch=not-foo seem to be harmless, therefore we
could try this approach instead. So, user must provide a list of arches not
supported by the backing pool, and we add --no-arch for each of them, letting wow
know we cannot run any tasks relevant just on those arches. It *still* may lead
to multiple recipeSets: e.g. if our backend supports x86_64, it supports i686
out of the box as well, and wow may split i686-only tasks to a separate box. But
this is not that harmful as the original issue.