openstack: guests must announce their architecture
The idea behind this is to tell openstack module what architecture it
provides, and let it pass this info to guests it creates. It is not yet
clear how all this arch stuff will work - openstack had historically
*just a single arch* approach, which we will need to leave. Will there
be special tenant for each arch, or will the instances be mixed in a
single tenant? Who knows. No matter what, openstack module will need to
know what architectures it provides, and it will need to propagate it to
its guests, hence this patch.