restraint-runner: template for restraint's results directory
Use the --results-directory-template to rename restraint's tmp*
directories to more usable names - tmp* is quite cryptic, not giving
user immediate info about what guest was used, what arch, what tests,
and so on. With this option, user has a power to make it more readable.

Not perfect, it's just a first attempt, therefore not compatible with
some options that would lead to using the same name for multiple
restraint runs. With some indices (e.g. order of the task/recipe)
available to the template, this issue should be fixable.