composetest: make it install
Given some limitations our setup.py has, it's necessary to statically
list module depedencies - they are not extracted from .moduleinfo files.
This will change in the future but now it's just the fact we have to live
with. composetest installs koji in a different way (from archive) while
we're used to using the system one - this may prove to be an issue when
we get to testing things when installed from rpm or in a container.
Let's see.