openstack, restraint-scheduler: instead of the exact image, tell provisioner what environemnt we need
restraint-scheduler is tightly coupled with openstack provisioner,
giving it an image - even when used with Beaker, restraint-scheduler
would try to look for an image. Should we add more backends (Docker,
AWS, ...), we'd need to teach scheduler how to translate an environment
(distro, arch, ...) to the actual image, which is not very flexible.
Hence offloading this to provisioners: scheduler(s) will ask for an
environment - in a way descibed by a protocol - and provisioners would
answer with guest(s).

openstack module now accepts this environment description but ignores it
- it calls `image` shared function (from guess-openstack-image) to get the
image. Which is good, scheduler doesn't care anymore, and provisioner can
use other modules to help with environment => image translation.

The long-term goal is simple - restraint + openstack/beaker. restraint
modules are still tight to openstack and some pieces of information are
simple expected, e.g. architectures or that the testing needed an
openstack image. This won't hold for Beaker.