bkr: new module, providing access to Beaker API and commands
Some Beaker-related actions seem to be shared among multiple modules -
it's not just `beaker` that needs to submit a job anymore. `beaker` name
is quite ambiguous in a sense it's our Beaker-backed pipeline worker and
at the same moment a module, talking to Beaker API, should carry this
name. Hence I opted for `bkr`, and this module should provide common
(low-level) actions like submitting a job or getting its results. All
this should be based on using `bkr` sub-commands.