restraint-scheduler: lower future check to check them as soon as they complete
Current code quit executor's context, which made the main thread wait
till all futures complete and then it resumed and moved on to checking
their results. By lowering this check into the context, we can peek at
each future as soon as it completes, without waiting for *all* of them
to finish. The main thread still can suspend but it would be awoken for
each newly completed future.