wow: allow caller of beaker_job_xml to override list of distros
`wow` module is smart. `wow` module is sometimes way too smart and tries to
guess bits we already know. Imagine this scenario:

- `restraint-scheduler` asks wow module to provide jobs to run
- `wow` replies with a bunch of XMLs
- `restraint-scheduler` prepares testing environments using these XMLs
- `restraint-scheduler` asks `beaker-provisioner` to provide a guest for
  each environment
- `beaker-provisioner` asks wow to provide simple job that reserves a
  guest of given architecture - **without forcing wow to use distro from
  the environment, giving it free hand to fill in this blank field**
- `wow` calls `distro` shared function to find distros, and for each
  distro one job is created

Imagine there are 2 distros, `A` and `B`, configured for a build target. On
the first call, `wow` would return 2 jobs reflecting this setup. On the second
call - when the specific distro is given by the testing environment
passed to `beaker-provisioner` - `wow` calls `distro` shared *again*, creating
2 jobs **again**, one for each distro, effectively multiplying number of guests:
first call runs produces 2 jobs, `A` and `B`; for testing environment `A` `wow`
would schedule `A` + `B` and for testing environment `B` `wow` would schedule `A`
and `B` once more.

Therefore we need a way to avoid using `distro` shared, forcing `wow` to
use the exact distro `beaker-provisioner` was told to use.