beaker-provisioner: use PHASE to inform wow that we're provisioning guests
This information is then used by wow's mappings wot add `--beah` option.
Since the default harness changed - and we're not yet able to deal with
multiple restraind instances on a single machine - we must use Beah when
provisioning guests. But only in that case - for everything else, we
stick with `--restraint` (which is now default anyway, and we are not
using `beaker-provisioner` to provision RHEL-8 boxes) since the
resulting XML would be given to `restraint`, and Beah and Restraint
views of a job XML may be incompatible.