rules-engine: "filter" functionality
A step between trivial evaluation of rules and complex instructions,
new `evaluate_filter` works like a `filter` built-in function for our
"instructions". It doesn't execute any commands at all, it simple
selects all entries whose ``rule`` evaluates to true-ish value.

The use case is a situation similar to our pattern maps - they map a
pattern to a string, often using build target as input. But we don't
have anythig to do that with rules more complex than "does the build
target match this pattern?" - this feature adds it, allowing us to
filter entries of lists based on complex rules.

For example, for OpenStack images, we have pattern maps. Using this
feature we'd have something a bit more verbose:

- rule: BUILD_TARGET.match('rhel-8\.\d+(?:\.\d+)?-.*')
  image: baseosci-RHEL-8.0-latest

- rule: BUILD_TARGET.match('(?:staging-|extras-|supp-|sjis-)?(?:rhel|RHEL)-6\.([56789]|10)-candidate')
  image: rhel-6.\1-server-x86_64-released

One thing missing is replacing references like \1 with proper values,
therefore this is not yet ready to replace our pattern maps.

The long-term goal is to get rid of BUILD_TARGET usage, replacing it
with testing environment descriptions.