publisher-umb-bus: reset error signal when we make successfull connection
The first attempt to send encountered few transport errors, but made a
successfull connection and send the message. But the error signal was
still set, and when container returned, the module checkd it and went
for another try. But the list of messages was empty already, therefore
this attempt #2 tried never got an opportunity to quit, because whether
there are any messages left to send is checked only after a message was
sent :)