Rename testing-environment.distro to .compose
"distro" seems to be a bit misleading, tied too much to Beaker, while
"compose" has quite well known meaning overall. We're not providing any
rigid definition of what values are allowed, that depends on modules who
set and consume it, we just feel "compose" is a better name for the
value.