beaker-provisioner, restraint: extra restraind instance and port it listens on
beaker-provisioner is using restraint to run "tests" that actualy take
care of provisioning the guests, therefore, this restraintd instance is
not available for running the actual tests because it is occupiend by
running the neverending reservesys task. We are able to start another
restraintd instance on a given port, and we can teach both provisioner
and restraint modules to accept this port and use it when running tests.

When using restraint module with OpenStack, nothing changes - we don't
use restraint to provision the guest, it is started to run tests, and
the port it listens to is the one restraint module knows as the default
one. Therefore no need to configure anything - it is possible now, but
not necessary.

With Beaker, the default restraintd is occupied, the second instance is
started and the pipeline needs to configure its restraint module to use
this second instance by changing default value of --restraintd-port.