beaker, test-schedule-runner-restraint: in results, keep order of tasks
When merging, task results collections were losing information about the
order in which tasks were executed. Fixing this, the order is preserved
- not perfectly, it is possible to fool the merging and produce final
result with the tasks in the incorrect order, but that requires user to
try to do exactly that by constructing suitable wow options. For common
folks, it should work well.

Implements PLATFORMCI-2265.