koji_fedora: the "has any artifacts at all?" check moved to the "user" side
When there are no artifact, koji module just gives up and raises an
error, despite the modules further in the pipeline may not even be
interested in the artifacts at all. It is not koji's place to decide
whether other modules would be allowed to run based on this check, since
the module cannot have any idea about what other modules are going to do
with the task. E.g. `koji jenkins brew-build-name` would fail for a task
without any artifacts despite the fact it is absolutely valid use case,
no module is actually *testing* artifacts, they are interested in the
task metadata only.

Patch moves the check to places that are interested in the artifacts,
i.e. modules that actually *run* testing and therefore would be harmed
shoudl there be no artifacts to, ehm, test.