wow: when no distro/arch combination is valid, warn user
May happen for example for a build with incomplete arches for target we
map to more than one distro. Imagine rhel-7.6-candidate build with just
an aarch64 arch - rhel-7.6-candidate is supposed to be tested on
RHEL-7.6 and RHEL-ALT-7.6, but since there's just aarch64, wow will fail
to prepare a plan for RHEL-7.6 since this distro does not support
aarch64.