koji: be smarter when deducing "complete" set of arches
Use a mapping, because we obviously cannot rely on Koji/Brew setting
arch_override option of the task.

Note that this is necessary to make peace between Koji/Brew and
workflow-tomorrow - test-scheduler bases pipeline does not suffer from
this problem because it inspects the task, and provisions necessary
architectures for thos present in the task, while legacy Beaker pipeline
and workflow-tomorrow can only *deny* a task via --no-arch option,
therefore they need to know what's the complete set, and add --no-arch
for architectures that are not present.