koji: task initializers, to bundle task and build IDs
Represents data we need to initialize a Koji task. A task ID would be enough, but, for some tasks,
we may need to override some data we'd otherwise get from Koji API.

The specific use case: container builds. Container build B1 was built by Brew task T1. Later,
there may be a rebuild of B1, thanks to change in the parent image, yilding B2. But: B2 would
point to T1! Thankfully, we can initialize with build ID (starting with B2 then), but because
our implementation would try to detect task behind B2 - which is, wrongly but officialy, T1 -
and use this task for initialization. Task instance would then try to detect build attached to
the task, which would be, according to API, B1... Therefore, we'd initialize with B2, but *nothing*
in our state would have any connection to B2, because the task behind B2 would be T1, and build
created by T1 would be B1.

To solve this trap, we need to preserve information about build after we reduce it to a task,
and when a task instance is initialized, we'd force this build to be the task is connected to.
Most of our code tries to use build when providing artifact attributes like NVR or component,
making it the information source number one.

Therefore task initializer, to give us a single package we could pass between involved functions.