koji_fedora: Do not use destination tag for latest relased
Previously, if destination tag existed, we used it by default in
resolving the latest released package. With introduction of RHEL8
gate tags, this is causing issues, as the destination tag is
rhel-8.1.0-gate, but the build taret is rhel-8.1.0-candidate, and the
former is the actual tag we want to resolve the latest packages from.

I believe we used the destination tag incorrectly from the start ...

This patch makes the latest_released a function, where the tag can
be overridden with the tag option in case somebody would like to
use the destination tag for resolving the latest released package.

The latest property stays, but obviously it does not use the destination
tag anymore.

The patch also makes the latest_released to return the first latest
tagged package in the list, if the NVR is different. This change is
again due to the introduction of the gate tags, where the latest
released package is actually the first tagged to the candidate tags.
Previously, as the destination tag and bulid target were the same,
we always wanted the second tagged package ...

Related: OSCI-364

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>