coldstore: artifacts location rendering
Restraint does it, restraint runner does it, STI wants to do it, future
script runner will need it - it's time to remove duplicity and provide
shared method for rendering pipeline artifact location.

Multiple modules have their own templates, to render a final location of
a pipeline artifact (e.g. restraint logs), but these templates are the
same: if there's coldstore, use it; if there's Jenkins, link to its
website; otherwise, add file:// and be done with it. Which is what we
want to do, we need a transparent method of switching *final* location
of the artifact. So, added to the coldstore module, and all relevant
modules were updated.