Use tasks' dispatch_id property when dispatching jobs
Our assumption that primary task's ID is perfectly fine for "artifact
ID" needed to dispatch the job does not hold anymore: to dispatch Docker
image testing, we need both the build ID and task ID to be passed to the
dispatched job. So let's give artifacts a chance to provide what they
think is relevant for dispatching.