wow: when no tasks were found, instead of exception return empty job
Leave the decision - and handling - of empty job (and schedule) to user
of wow module.

When scheduling jobs for multiple distros (D1, D2) - while there are
tasks relevant for D1, should there be no tasks relevant for D2, wow
module would raise an exception because of D2 yielding an empty job,
killing the whole pipeline. It doesn't track non-empty job based on D1,
wow's user is the one supposed to make a call and continue/quit.