ansible, guest-setup: focus on a single guest
We're dropping support for parallelization built-in into Ansible in
favor of higher-level approach with futures: if you need to run
playbooks on multiple guests, spawn futures. Ansible and
guest-setup-like modules now deal with a single guest only, and this
gives us chance to simplify logging. Each guest-setup-like module now
works with a given directory - usually guest-specific - and we have
a helper for translation between local paths and user-visible locations.
Everything is now nicely contained.