copr: fix JSONDecodeError
Sometimes the copr module gets an unsuccesful response from copr. The response doesn't raise an error in the `get()` method, but whenever `json()` tries to parse the response - it fails. The patch resolves the issue with moving raising of the error to the same level, where `json()` method is executing.