jenkins: "lazy" wrappers over jenkinsapi job and build
We're asking jenkinsapi to give us "job" representation, but to create
such object, jenkinsapi tries to fetch info for all builds of the given
job. And that's something we're not interested in, not at all - all we
need is job and build IDs and some way to invoke job.

So, let's avoid asking jenkinsapi for job representation as much as
possible - simple objects to carry job/build IDs, and our own simple
"invoke job" method (since invoking a job is just a matter of submitting
a form).