ansible: fix check of user-set debug callback
The ANSIBLE_STDOUT_CALLBACK variable may be set in the environment, and
our attempt to avoid changing it when enabling prettier Ansible output
would fail despite the fact that the variable might be set to "debug",
the very same value we'd set.

So, out check whether ANSIBLE_STDOUT_CALLBACK exists too hard - let's
check whether it exists and it's set to something different than
"debug". If it exists and it's set to "debug", we will add our "-v"
option to get nicely formatted output.

This also revealed a problem with tests - they pass when this variable
is set, but fail otherwise. Fixing them as well.