guest-setup: support pre-, post- and artifact installation stages
We need to perform actions before the installation (we already do),
during the installation (we already do), and after the installation -
this we don't do, and we're solving this task in a weird way by
employing wow.

Patch adds different stages, calling the current behavior
pre-artifact-installation stage. This way we can add more action to this
"setup guest before the testing" pipeline step.