guess-environment: has to accept multiple maps
We're trying to match destination tag first, then build target, but we
were using the same map for both which yields unexpected results. Of
course it does because the map is tailored for matching build targets,
not destination tags. Therefore we need a way to specify map for each
"test".