beaker-jobwatch: instead of hardcoding options, read them from module config
--skip-broken-machines is perfectly reasonable, valid and mostly useful
option. However, it breaks the workflow when the beaker jobs can use
only a very small set of machines, e.g. 1. When the job fails and gets
rescheduled, --skip-broken-machines add explicit rule to the XML to
avoid the machine. These rules cumulate and may disallow all possible
machines, killing rescheduled jobs because there'd be no HW matching the
request. So, when users have to use small set of machines, they get a
chance to remove --skip-broken-machines from the beaker-jobwatch options.
The rescheduled beaker jobs may crash again, but it's a chance they deserve.