koji: allow failed tasks to be processed
1. while testing containers, we encountered Brew tasks that were marked
as "failed" yet yielded valid builds we need to test. Since our
pipelines use explicit build ID to begin with, we don't really need the
task. The only block was the task's state check which did not allow
failed/canceled tasks to continue. This patch adds config option that
would allow us to accept these tasks when needed.

2. did I say that's the only block? I lied. Brew class also inspected
tasks' result property, which usualy carries corresponding build ID. But
we already have a build ID to use, set explicitly by "user", so we don't
need to even read the result - that's the other change, a genuine issue
in our code. Not really a bug, but a "waste" of CPU time. It was
revealed by accepting the failed tasks, their results are broken as
well. But we don't have to inspect their results, right?